setup eslint & prettier
