+++ .dockerignore
@@ -2,3 +2,4 @@ dist/
 .git
 .gitignore
 *.md
+.DS_Store
