+++ .eslintrc
@@ -1,0 +1,60 @@
+{
+    "extends": [
+        "eslint:recommended",
+        "plugin:@typescript-eslint/recommended",
+        "plugin:@typescript-eslint/recommended-requiring-type-checking",
+        "next/core-web-vitals",
+        "prettier"
+    ],
+    "parser": "@typescript-eslint/parser",
+    "parserOptions": {
+        "project": "./tsconfig.json"
+    },
+    "plugins": ["@typescript-eslint"],
+    "root": true,
+    "rules": {
+        "import/order": [
+            "error",
+            {
+                "groups": ["builtin", "external", "internal", ["parent", "sibling"], "object", "type", "index"],
+                "newlines-between": "always",
+                "pathGroupsExcludedImportTypes": ["builtin"],
+                "pathGroups": [
+                    {
+                        "pattern": "@/utils/**",
+                        "group": "internal",
+                        "position": "before"
+                    },
+                    {
+                        "pattern": "@/libs/**",
+                        "group": "internal",
+                        "position": "before"
+                    },
+                    {
+                        "pattern": "@/hooks/**",
+                        "group": "internal",
+                        "position": "before"
+                    },
+                    {
+                        "pattern": "@/components/**",
+                        "group": "internal",
+                        "position": "before"
+                    },
+                    {
+                        "pattern": "@/const/**",
+                        "group": "internal",
+                        "position": "before"
+                    },
+                    {
+                        "pattern": "@/types/**",
+                        "group": "internal",
+                        "position": "before"
+                    }
+                ],
+                "alphabetize": {
+                    "order": "asc"
+                }
+            }
+        ]
+    }
+}
