+++ .prettierignore
@@ -1,0 +1,27 @@
+# Ignore artifacts:
+build
+coverage
+yarn.lock
+package-lock.json
+.DS_Store
+
+# dotfile
+.env*
+
+# markdown
+*.md
+
+# next.js
+/.next/
+/out/
+
+# production
+/build
+
+# yarn
+/.yarn/
+
+# materials
+public/
+*.ico
+*.svg
