+++ .vscode/settings.json
@@ -1,0 +1,9 @@
+{
+    "files.insertFinalNewline": true,
+    "prettier.configPath": ".prettierrc",
+    "editor.defaultFormatter": "esbenp.prettier-vscode",
+    "editor.formatOnSave": true,
+    "editor.codeActionsOnSave": {
+        "source.fixAll.eslint": "always",
+    }
+}
