+++ postcss.config.js
@@ -1,8 +1,8 @@
 /** @type {import('postcss-load-config').Config} */
 const config = {
-  plugins: {
-    tailwindcss: {},
-  },
+    plugins: {
+        tailwindcss: {},
+    },
 };
 
 export default config;
